+++ src/OF_WAR.cpp
@@ -663,6 +663,8 @@ static void i_disp_build_button(ButtonCustom *button, int repaintBody)
 			str += "<";
 
 		font_bible.put( x1+BUILD_UNIT_NAME_OFFSET_X, y1+BUILD_UNIT_NAME_OFFSET_Y, str );
+
+		Ambition::Draw::buttonCost(font_bible, unitInfo->build_cost, x1, y1 + BUILD_UNIT_NAME_OFFSET_Y, x2 - 48, y2, -1);
 	}
 
 	// display small button